feat: extract reason-emoji map and shared types to stars modules
Move the REASON_EMOJI lookup table and formatRelativeTime helper into
web/src/utils/stars.ts, and the Balance, Transaction, streak, and journey
interfaces into a new web/src/types/stars.ts (mirroring types/leaderboard.ts).
Stars.tsx now imports these instead of declaring them inline. Pure refactor
with no behavior change.

Bead: Hytte-cugq2
